LocateControl, rename components helper files, update readme wip
