+++ README.md
@@ -1,12 +1,22 @@
 # @radiofrance/svelte-leaflet
 
-A library that wraps leaflet classes in domless/renderless svelte components.
+Simple Leaflet maps for your Svelte projects.
+
+## Features
+
+- Works in SSR pages
+- TypeScript : events, props and options are fully typed
+- Simple event usage with the `oneventname` syntax
+  - ![](./static/map-types.png)
+  - `event.sourceTarget` is correctly typed
+- Fully reactive : update props and options, the components will update
+- Customize `DivIcon`, `Popup` and more with Svelte components or HTML instead of strings
+- Single dependency : no `leaflet` or `@types/leaflet` needed
 
 ## Compatibility
 
-Version 1.0.0 and above are compatible with Svelte 5.
-
-For svelte 3 and 4, use version 0.1.x (not actively maintained).
+- Version 1.0.0 and above are compatible with Svelte 5.
+- For Svelte 3 and 4, use version 0.1.x (not actively maintained).
 
 ## Install
 
@@ -14,82 +24,57 @@ For svelte 3 and 4, use version 0.1.x (not actively maintained).
 
 ## Components
 
-### Map
+### UI Layers
 
-Renders a map Leaflet Map. The Map will take up 100% of its container's height and width.
+- ✅ Marker
+- DivOverlay
+- ✅ Popup
+- Tooltip
 
-#### Attributes
+### Raster Layers
 
-| Attribute     | Type                                                          | Default                                            | Notes                                                                              |
-| ------------- | ------------------------------------------------------------- | -------------------------------------------------- | ---------------------------------------------------------------------------------- |
-| `options`     | [MapOptions](https://leafletjs.com/reference.html#map-option) | `{}`                                               | Map options                                                                        |
-| `tilesUrl`    | string                                                        | `'https://tile.openstreetmap.org/{z}/{x}/{y}.png'` | more free tile services can be found at https://alexurquhart.github.io/free-tiles/ |
-| `attribution` | string                                                        | `'©OpenStreetMap'` _(link to openstreetmap)_      |                                                                                    |
+- TileLayer
+- TileLayer.WMS
+- ImageOverlay
+- VideoOverlay
 
-#### Events
+### Vector Layers
 
-All events are forwarded from the Map class, see the leaflet documentation for more information.
+- Path
+- ✅ Polyline
+- ✅ Polygon
+- Rectangle
+- Circle
+- CircleMarker
+- SVGOverlay
+- SVG
+- Canvas
 
-- https://leafletjs.com/reference.html#map-event
+### Other Layers
 
-Example:
+- LayerGroup
+- FeatureGroup
+- GeoJSON
+- GridLayer
 
-```svelte
-<Map {options} onclick={(e) => console.log(e.detail.latlng)} />
-```
+### Basic Types
 
-### Marker
+- LatLng
+- LatLngBounds
+- Point
+- Bounds
+- ✅ Icon
+- ✅ DivIcon
 
-Add a marker to the map.
+### Controls
 
-- Can be used as a child of `<Map>` or `<MarkerClusterGroup>`
-- A `<Popup>` component can be passed as the Marker child to display a popup when the marker is clicked.
-- To use a custom icon, pass a `Icon` or a `DivIcon` component as the marker child.
+- ✅ Zoom
+- Attribution
+- Layers
+- Scale
 
-#### Attributes
+### Extra
 
-| Name      | Type                                                                | Default    | Notes                                 |
-| --------- | ------------------------------------------------------------------- | ---------- | ------------------------------------- |
-| `latlng`  | [LatLng](https://leafletjs.com/reference.html#latlng)               | _required_ | position of the marker                |
-| `options` | [MarkerOptions](https://leafletjs.com/reference.html#marker-option) | `{}`       | options to pass to the leaflet marker |
-
-#### Events
-
-All events are forwarded from the Marker class, see the leaflet documentation for more information.
-
-- https://leafletjs.com/reference.html#marker-event
-
-### Popup
-
-`<Popup>` children will be displayed when the marker is clicked.
-
-- Use as a `<Marker>` child.
-
-#### Attributes
-
-| Name      | Type                                                              | Default | Notes                                |
-| --------- | ----------------------------------------------------------------- | ------- | ------------------------------------ |
-| `options` | [PopupOptions](https://leafletjs.com/reference.html#popup-option) | `{}`    | options to pass to the leaflet popup |
-
-### MarkerClusterGroup
-
-Enables clustering of child markers
-
-- Use as a `<Map>` child
-
-#### Attributes
-
-| Name      | Type                                                                                                                                   | Default | Notes                                                                                                      |
-| --------- | -------------------------------------------------------------------------------------------------------------------------------------- | ------- | ---------------------------------------------------------------------------------------------------------- |
-| `options` | [MarkerClusterGroupOptions](https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/leaflet.markercluster/index.d.ts#L36) | `{}`    | [MarkerClusterGroup options](https://github.com/leaflet/Leaflet.markercluster?tab=readme-ov-file#defaults) |
-
-### Polyline
-
-Allows to draw lines on the map
-
-#### Attributes
-
-| Name      | Type                                                                    | Default    | Notes                                   |
-| --------- | ----------------------------------------------------------------------- | ---------- | --------------------------------------- |
-| `latlngs` | [LatLng[]](https://leafletjs.com/reference.html#latlng)                 | _required_ | array of points to draw the line        |
-| `options` | [PolylineOptions](https://leafletjs.com/reference.html#polyline-option) | `{}`       | options to pass to the leaflet polyline |
+- ✅ MarkerClusterGroup
+- ✅ LocateControl
+  - ![locate control](./static//locate-control.png)
